app/requirements.txt
flask
gunicorn
requests
azure.data.tables
